ImprovSurvey/requirements.txt
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.30.0
dnspython==2.2.1
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
Pillow==9.2.0
pymongo==4.2.0
python-arptable==0.0.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2